Add cluster role bindings required to make helm work
Adding the Tiller service account which is required for Tiller to work,
also adding a bit of a hack to give the cluster's default service
account (as created by GKE) cluster admin priviledges. The latter is
required b/c we are running helm from inside the cluster - eventually we
will be doing this with our own service account and won't need this.

Co-authored-by: Christie Wilson <bobcatfish@gmail.com>